requirements.txt
absl-py==2.2.2
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==25.3.0
av==12.0.0
blobfile==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
cmake==4.0.0
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
DALL-E==0.1
decord==0.6.0
diffusers==0.27.0
einops==0.8.1
exceptiongroup==1.2.2
filelock==3.17.0
fire==0.7.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2025.2.0
future==1.0.0
grpcio==1.71.0
huggingface-hub==0.25.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.1.0
Jinja2==3.1.5
kiwisolver==1.4.8
lightning-lite==1.8.0
lightning-utilities==0.3.0
lit==18.1.8
lxml==5.3.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
multidict==6.4.3
mypy==1.15.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.24.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.0
onnxruntime==1.20.1
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pip==23.3
pluggy==1.5.0
propcache==0.3.1
protobuf==3.20.1
psutil==7.0.0
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pyDeprecate==0.3.2
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==1.6.5
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.2
setuptools==75.8.0
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.4
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
test_tube==0.7.5
tokenizers==0.21.0
tomli==2.2.1
torch==1.13.0
torchaudio==2.1.2+cu121
torchmetrics==0.11.4
torchvision==0.14.0
tqdm==4.67.1
transformers==4.48.3
triton==2.0.0
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.0
zipp==3.21.0
